+++ CHANGELOG.rst
@@ -23,12 +23,12 @@ number of the code change for that issue.  These PRs can be viewed at:
 
 - Added code to handle deprecations in photutils v3.0.0. [#2089]
 
+- Replaced stregion with astropy-regions in mapreg.py, a module for reading and 
+  parsing DS9 region files used for tweakreg. [#2125]
+
 
 3.11.0 (28-Apr-2026)
 ====================
-
-- Replaced stregion with astropy-regions in mapreg.py, a module for reading and 
-  parsing DS9 region files used for tweakreg. [#2125]
 
 - Deprecated the hapcut_utils module. [#2103]