+++ CHANGELOG.rst
@@ -23,7 +23,7 @@ number of the code change for that issue.  These PRs can be viewed at:
 
 - Added code to handle deprecations in photutils v3.0.0. [#2089]
 
-- Deprecated the teal helper "run" functions as well as the editpars 
+- Deprecated the TEAL helper "run" functions as well as the editpars 
   parameters in base drizzlepac functions. [#2152]